requirements.txt
google-generativeai
beautifulsoup4
chromadb
pypdf
numpy
python-dotenv
Pillow
prompt_toolkit
requests
textual
